requirements.txt
Django>=5.0
djangorestframework>=3.14
django-environ>=0.11
psycopg2-binary>=2.9
httpx>=0.26
gunicorn>=21.0
djangorestframework-simplejwt>=5.3
django-filter>=23.0
pytest>=7.0
pytest-cov>=4.0
pytest-django>=4.5
factory-boy>=3.3
flake8>=6.0
black>=23.0
beautifulsoup4>=4.12
django-allauth>=0.63
